+++ README.md
@@ -90,7 +90,7 @@ These GitHub repositories provide supplementary resources for Rush Stack:
 | [/webpack/webpack-plugin-utilities](./webpack/webpack-plugin-utilities/) | [![npm version](https://badge.fury.io/js/%40rushstack%2Fwebpack-plugin-utilities.svg)](https://badge.fury.io/js/%40rushstack%2Fwebpack-plugin-utilities) | [changelog](./webpack/webpack-plugin-utilities/CHANGELOG.md) | [@rushstack/webpack-plugin-utilities](https://www.npmjs.com/package/@rushstack/webpack-plugin-utilities) |
 | [/webpack/webpack4-localization-plugin](./webpack/webpack4-localization-plugin/) | [![npm version](https://badge.fury.io/js/%40rushstack%2Fwebpack4-localization-plugin.svg)](https://badge.fury.io/js/%40rushstack%2Fwebpack4-localization-plugin) | [changelog](./webpack/webpack4-localization-plugin/CHANGELOG.md) | [@rushstack/webpack4-localization-plugin](https://www.npmjs.com/package/@rushstack/webpack4-localization-plugin) |
 | [/webpack/webpack4-module-minifier-plugin](./webpack/webpack4-module-minifier-plugin/) | [![npm version](https://badge.fury.io/js/%40rushstack%2Fwebpack4-module-minifier-plugin.svg)](https://badge.fury.io/js/%40rushstack%2Fwebpack4-module-minifier-plugin) | [changelog](./webpack/webpack4-module-minifier-plugin/CHANGELOG.md) | [@rushstack/webpack4-module-minifier-plugin](https://www.npmjs.com/package/@rushstack/webpack4-module-minifier-plugin) |
-| [/webpack/webpack5-loader-load-themed-styles](./webpack/webpack5-loader-load-themed-styles/) | [![npm version](https://badge.fury.io/js/%40microsoft%2Fwebpack5-loader-load-themed-styles.svg)](https://badge.fury.io/js/%40microsoft%2Fwebpack5-loader-load-themed-styles) | [changelog](./webpack/webpack5-loader-load-themed-styles/CHANGELOG.md) | [@microsoft/webpack5-loader-load-themed-styles](https://www.npmjs.com/package/@microsoft/webpack5-loader-load-themed-styles) |
+| [/webpack/webpack5-loader-load-themed-styles](./webpack/webpack5-loader-load-themed-styles/) | [![npm version](https://badge.fury.io/js/%40microsoft%2Fwebpack5-loader-load-themed-styles.svg)](https://badge.fury.io/js/%40microsoft%2Fwebpack5-loader-load-themed-styles) | [changelog](./webpack/webpack5-loader-load-themed-styles/CHANGELOG.md) | [@microsoft/webpack5-load-themed-styles-loader](https://www.npmjs.com/package/@microsoft/webpack5-load-themed-styles-loader) |
 | [/webpack/webpack5-localization-plugin](./webpack/webpack5-localization-plugin/) | [![npm version](https://badge.fury.io/js/%40rushstack%2Fwebpack5-localization-plugin.svg)](https://badge.fury.io/js/%40rushstack%2Fwebpack5-localization-plugin) | [changelog](./webpack/webpack5-localization-plugin/CHANGELOG.md) | [@rushstack/webpack5-localization-plugin](https://www.npmjs.com/package/@rushstack/webpack5-localization-plugin) |
 | [/webpack/webpack5-module-minifier-plugin](./webpack/webpack5-module-minifier-plugin/) | [![npm version](https://badge.fury.io/js/%40rushstack%2Fwebpack5-module-minifier-plugin.svg)](https://badge.fury.io/js/%40rushstack%2Fwebpack5-module-minifier-plugin) | [changelog](./webpack/webpack5-module-minifier-plugin/CHANGELOG.md) | [@rushstack/webpack5-module-minifier-plugin](https://www.npmjs.com/package/@rushstack/webpack5-module-minifier-plugin) |
