--- build-tests/api-documenter-scenarios/etc/inheritedMembers/api-documenter-scenarios.api.md
@@ -1,89 +1,0 @@
-## API Report File for "api-documenter-scenarios"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-import { Extractor } from '@microsoft/api-extractor';
-
-// @public (undocumented)
-export class Class1 extends Class2<number> {
-    fourthProp: number;
-    secondProp: boolean;
-    someOverload(x: boolean | string): void;
-}
-
-// @public (undocumented)
-export class Class2<T> extends Namespace1.Class3 {
-    secondProp: boolean | string;
-    someMethod(x: boolean): void;
-    thirdProp: T;
-}
-
-// @public
-export const ClassLikeVariable: {
-    new (): {
-        someProp: number;
-    };
-};
-
-// Warning: (ae-forgotten-export) The symbol "ExtendsAnonymousClass_base" needs to be exported by the entry point index.d.ts
-//
-// @public
-export class ExtendsAnonymousClass extends ExtendsAnonymousClass_base {
-}
-
-// @public
-export class ExtendsClassFromAnotherPackage extends Extractor {
-}
-
-// @public
-export class ExtendsClassLikeVariable extends ClassLikeVariable {
-}
-
-// Warning: (ae-forgotten-export) The symbol "UnexportedClass" needs to be exported by the entry point index.d.ts
-//
-// @public
-export class ExtendsUnexportedClass extends UnexportedClass {
-}
-
-// @public
-export interface IExtendsInterfaceLikeTypeAlias extends IInterfaceLikeTypeAlias, IInterface1 {
-}
-
-// @public
-export interface IExtendsMultipleInterfaces extends IInterface1, IInterface2 {
-    secondProp: boolean;
-    thirdProp: string;
-}
-
-// @public (undocumented)
-export interface IInterface1 {
-    secondProp: boolean | string;
-    someProp: number;
-}
-
-// @public (undocumented)
-export interface IInterface2 {
-    someProp: number;
-}
-
-// @public
-export type IInterfaceLikeTypeAlias = {
-    someProp: number;
-};
-
-// @public (undocumented)
-export namespace Namespace1 {
-    // (undocumented)
-    export class Class3 {
-        someMethod(x: boolean | string): void;
-        someOverload(x: boolean): void;
-        someOverload(x: string): void;
-        someProp: number;
-    }
-}
-
-// (No @packageDocumentation comment for this package)
-
-```
