--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.class1.fourthprop.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [Class1](./api-documenter-scenarios.class1.md) &gt; [fourthProp](./api-documenter-scenarios.class1.fourthprop.md)
-
-## Class1.fourthProp property
-
-A fourth prop
-
-<b>Signature:</b>
-
-```typescript
-fourthProp: number;
-```
