--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.class1.md
@@ -1,30 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [Class1](./api-documenter-scenarios.class1.md)
-
-## Class1 class
-
-
-<b>Signature:</b>
-
-```typescript
-export declare class Class1 extends Class2<number> 
-```
-<b>Extends:</b> [Class2](./api-documenter-scenarios.class2.md)<!-- -->&lt;number&gt;
-
-## Properties
-
-|  Property | Modifiers | Type | Description |
-|  --- | --- | --- | --- |
-|  [fourthProp](./api-documenter-scenarios.class1.fourthprop.md) |  | number | A fourth prop |
-|  [secondProp](./api-documenter-scenarios.class1.secondprop.md) |  | boolean | A second prop. Overrides <code>Class2.secondProp</code>. |
-|  [someProp](./api-documenter-scenarios.namespace1.class3.someprop.md) |  | number | <p>Some prop.</p><p>(Inherited from [Class3](./api-documenter-scenarios.namespace1.class3.md)<!-- -->)</p> |
-|  [thirdProp](./api-documenter-scenarios.class2.thirdprop.md) |  | T | <p>A third prop.</p><p>(Inherited from [Class2](./api-documenter-scenarios.class2.md)<!-- -->)</p> |
-
-## Methods
-
-|  Method | Modifiers | Description |
-|  --- | --- | --- |
-|  [someMethod(x)](./api-documenter-scenarios.class2.somemethod.md) |  | <p>Some method. Overrides <code>Class3.someMethod</code>.</p><p>(Inherited from [Class2](./api-documenter-scenarios.class2.md)<!-- -->)</p> |
-|  [someOverload(x)](./api-documenter-scenarios.class1.someoverload.md) |  | Some overload. Overrides <code>Class3.someOverload</code>. |
-
