--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.class1.secondprop.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [Class1](./api-documenter-scenarios.class1.md) &gt; [secondProp](./api-documenter-scenarios.class1.secondprop.md)
-
-## Class1.secondProp property
-
-A second prop. Overrides `Class2.secondProp`<!-- -->.
-
-<b>Signature:</b>
-
-```typescript
-secondProp: boolean;
-```
