--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.class1.someoverload.md
@@ -1,24 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [Class1](./api-documenter-scenarios.class1.md) &gt; [someOverload](./api-documenter-scenarios.class1.someoverload.md)
-
-## Class1.someOverload() method
-
-Some overload. Overrides `Class3.someOverload`<!-- -->.
-
-<b>Signature:</b>
-
-```typescript
-someOverload(x: boolean | string): void;
-```
-
-## Parameters
-
-|  Parameter | Type | Description |
-|  --- | --- | --- |
-|  x | boolean \| string |  |
-
-<b>Returns:</b>
-
-void
-
