--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.class2.md
@@ -1,30 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [Class2](./api-documenter-scenarios.class2.md)
-
-## Class2 class
-
-
-<b>Signature:</b>
-
-```typescript
-export declare class Class2<T> extends Namespace1.Class3 
-```
-<b>Extends:</b> [Namespace1.Class3](./api-documenter-scenarios.namespace1.class3.md)
-
-## Properties
-
-|  Property | Modifiers | Type | Description |
-|  --- | --- | --- | --- |
-|  [secondProp](./api-documenter-scenarios.class2.secondprop.md) |  | boolean \| string | A second prop. |
-|  [someProp](./api-documenter-scenarios.namespace1.class3.someprop.md) |  | number | <p>Some prop.</p><p>(Inherited from [Class3](./api-documenter-scenarios.namespace1.class3.md)<!-- -->)</p> |
-|  [thirdProp](./api-documenter-scenarios.class2.thirdprop.md) |  | T | A third prop. |
-
-## Methods
-
-|  Method | Modifiers | Description |
-|  --- | --- | --- |
-|  [someMethod(x)](./api-documenter-scenarios.class2.somemethod.md) |  | Some method. Overrides <code>Class3.someMethod</code>. |
-|  [someOverload(x)](./api-documenter-scenarios.namespace1.class3.someoverload.md) |  | <p>Some overload.</p><p>(Inherited from [Class3](./api-documenter-scenarios.namespace1.class3.md)<!-- -->)</p> |
-|  [someOverload(x)](./api-documenter-scenarios.namespace1.class3.someoverload_1.md) |  | <p>Some overload.</p><p>(Inherited from [Class3](./api-documenter-scenarios.namespace1.class3.md)<!-- -->)</p> |
-
