--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.class2.somemethod.md
@@ -1,24 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [Class2](./api-documenter-scenarios.class2.md) &gt; [someMethod](./api-documenter-scenarios.class2.somemethod.md)
-
-## Class2.someMethod() method
-
-Some method. Overrides `Class3.someMethod`<!-- -->.
-
-<b>Signature:</b>
-
-```typescript
-someMethod(x: boolean): void;
-```
-
-## Parameters
-
-|  Parameter | Type | Description |
-|  --- | --- | --- |
-|  x | boolean |  |
-
-<b>Returns:</b>
-
-void
-
