--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.class2.thirdprop.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [Class2](./api-documenter-scenarios.class2.md) &gt; [thirdProp](./api-documenter-scenarios.class2.thirdprop.md)
-
-## Class2.thirdProp property
-
-A third prop.
-
-<b>Signature:</b>
-
-```typescript
-thirdProp: T;
-```
