--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.classlikevariable.md
@@ -1,17 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [ClassLikeVariable](./api-documenter-scenarios.classlikevariable.md)
-
-## ClassLikeVariable variable
-
-Some class-like variable.
-
-<b>Signature:</b>
-
-```typescript
-ClassLikeVariable: {
-    new (): {
-        someProp: number;
-    };
-}
-```
