--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.extendsanonymousclass.md
@@ -1,17 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [ExtendsAnonymousClass](./api-documenter-scenarios.extendsanonymousclass.md)
-
-## ExtendsAnonymousClass class
-
-Some class that extends an anonymous class.
-
-<b>Signature:</b>
-
-```typescript
-export declare class ExtendsAnonymousClass extends ExtendsAnonymousClass_base 
-```
-<b>Extends:</b> ExtendsAnonymousClass\_base
-
-<i>(Some inherited members may not be shown because they are not represented in the documentation.)</i>
-
