--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.extendsclassfromanotherpackage.md
@@ -1,17 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [ExtendsClassFromAnotherPackage](./api-documenter-scenarios.extendsclassfromanotherpackage.md)
-
-## ExtendsClassFromAnotherPackage class
-
-Some class that extends a class from another package. This base class is not in any API doc model.
-
-<b>Signature:</b>
-
-```typescript
-export declare class ExtendsClassFromAnotherPackage extends Extractor 
-```
-<b>Extends:</b> Extractor
-
-<i>(Some inherited members may not be shown because they are not represented in the documentation.)</i>
-
