--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.extendsclasslikevariable.md
@@ -1,17 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [ExtendsClassLikeVariable](./api-documenter-scenarios.extendsclasslikevariable.md)
-
-## ExtendsClassLikeVariable class
-
-Some class that extends a class-like variable.
-
-<b>Signature:</b>
-
-```typescript
-export declare class ExtendsClassLikeVariable extends ClassLikeVariable 
-```
-<b>Extends:</b> ClassLikeVariable
-
-<i>(Some inherited members may not be shown because they are not represented in the documentation.)</i>
-
