--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.extendsunexportedclass.md
@@ -1,17 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [ExtendsUnexportedClass](./api-documenter-scenarios.extendsunexportedclass.md)
-
-## ExtendsUnexportedClass class
-
-Some class that extends an unexported class.
-
-<b>Signature:</b>
-
-```typescript
-export declare class ExtendsUnexportedClass extends UnexportedClass 
-```
-<b>Extends:</b> UnexportedClass
-
-<i>(Some inherited members may not be shown because they are not represented in the documentation.)</i>
-
