--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.iextendsmultipleinterfaces.md
@@ -1,23 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [IExtendsMultipleInterfaces](./api-documenter-scenarios.iextendsmultipleinterfaces.md)
-
-## IExtendsMultipleInterfaces interface
-
-Some interface that extends multiple interfaces.
-
-<b>Signature:</b>
-
-```typescript
-export interface IExtendsMultipleInterfaces extends IInterface1, IInterface2 
-```
-<b>Extends:</b> [IInterface1](./api-documenter-scenarios.iinterface1.md)<!-- -->, [IInterface2](./api-documenter-scenarios.iinterface2.md)
-
-## Properties
-
-|  Property | Modifiers | Type | Description |
-|  --- | --- | --- | --- |
-|  [secondProp](./api-documenter-scenarios.iextendsmultipleinterfaces.secondprop.md) |  | boolean | A second prop. Overrides <code>IInterface1.someProp</code>. |
-|  [someProp](./api-documenter-scenarios.iinterface1.someprop.md) |  | number | <p>Some prop.</p><p>(Inherited from [IInterface1](./api-documenter-scenarios.iinterface1.md)<!-- -->)</p> |
-|  [thirdProp](./api-documenter-scenarios.iextendsmultipleinterfaces.thirdprop.md) |  | string | A third prop. |
-
