--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.iextendsmultipleinterfaces.secondprop.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [IExtendsMultipleInterfaces](./api-documenter-scenarios.iextendsmultipleinterfaces.md) &gt; [secondProp](./api-documenter-scenarios.iextendsmultipleinterfaces.secondprop.md)
-
-## IExtendsMultipleInterfaces.secondProp property
-
-A second prop. Overrides `IInterface1.someProp`<!-- -->.
-
-<b>Signature:</b>
-
-```typescript
-secondProp: boolean;
-```
