--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.iextendsmultipleinterfaces.thirdprop.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [IExtendsMultipleInterfaces](./api-documenter-scenarios.iextendsmultipleinterfaces.md) &gt; [thirdProp](./api-documenter-scenarios.iextendsmultipleinterfaces.thirdprop.md)
-
-## IExtendsMultipleInterfaces.thirdProp property
-
-A third prop.
-
-<b>Signature:</b>
-
-```typescript
-thirdProp: string;
-```
