--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.iinterface1.md
@@ -1,20 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [IInterface1](./api-documenter-scenarios.iinterface1.md)
-
-## IInterface1 interface
-
-
-<b>Signature:</b>
-
-```typescript
-export interface IInterface1 
-```
-
-## Properties
-
-|  Property | Modifiers | Type | Description |
-|  --- | --- | --- | --- |
-|  [secondProp](./api-documenter-scenarios.iinterface1.secondprop.md) |  | boolean \| string | A second prop. |
-|  [someProp](./api-documenter-scenarios.iinterface1.someprop.md) |  | number | Some prop. |
-
