--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.iinterface1.secondprop.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [IInterface1](./api-documenter-scenarios.iinterface1.md) &gt; [secondProp](./api-documenter-scenarios.iinterface1.secondprop.md)
-
-## IInterface1.secondProp property
-
-A second prop.
-
-<b>Signature:</b>
-
-```typescript
-secondProp: boolean | string;
-```
