--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.iinterface2.md
@@ -1,19 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [IInterface2](./api-documenter-scenarios.iinterface2.md)
-
-## IInterface2 interface
-
-
-<b>Signature:</b>
-
-```typescript
-export interface IInterface2 
-```
-
-## Properties
-
-|  Property | Modifiers | Type | Description |
-|  --- | --- | --- | --- |
-|  [someProp](./api-documenter-scenarios.iinterface2.someprop.md) |  | number | Some prop. |
-
