--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.iinterface2.someprop.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [IInterface2](./api-documenter-scenarios.iinterface2.md) &gt; [someProp](./api-documenter-scenarios.iinterface2.someprop.md)
-
-## IInterface2.someProp property
-
-Some prop.
-
-<b>Signature:</b>
-
-```typescript
-someProp: number;
-```
