--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.iinterfaceliketypealias.md
@@ -1,15 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [IInterfaceLikeTypeAlias](./api-documenter-scenarios.iinterfaceliketypealias.md)
-
-## IInterfaceLikeTypeAlias type
-
-Some interface-like type alias.
-
-<b>Signature:</b>
-
-```typescript
-export declare type IInterfaceLikeTypeAlias = {
-    someProp: number;
-};
-```
