--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.md
@@ -1,44 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md)
-
-## api-documenter-scenarios package
-
-## Classes
-
-|  Class | Description |
-|  --- | --- |
-|  [Class1](./api-documenter-scenarios.class1.md) |  |
-|  [Class2](./api-documenter-scenarios.class2.md) |  |
-|  [ExtendsAnonymousClass](./api-documenter-scenarios.extendsanonymousclass.md) | Some class that extends an anonymous class. |
-|  [ExtendsClassFromAnotherPackage](./api-documenter-scenarios.extendsclassfromanotherpackage.md) | Some class that extends a class from another package. This base class is not in any API doc model. |
-|  [ExtendsClassLikeVariable](./api-documenter-scenarios.extendsclasslikevariable.md) | Some class that extends a class-like variable. |
-|  [ExtendsUnexportedClass](./api-documenter-scenarios.extendsunexportedclass.md) | Some class that extends an unexported class. |
-
-## Interfaces
-
-|  Interface | Description |
-|  --- | --- |
-|  [IExtendsInterfaceLikeTypeAlias](./api-documenter-scenarios.iextendsinterfaceliketypealias.md) | Some interface that extends an interface-like type alias as well as another interface. |
-|  [IExtendsMultipleInterfaces](./api-documenter-scenarios.iextendsmultipleinterfaces.md) | Some interface that extends multiple interfaces. |
-|  [IInterface1](./api-documenter-scenarios.iinterface1.md) |  |
-|  [IInterface2](./api-documenter-scenarios.iinterface2.md) |  |
-
-## Namespaces
-
-|  Namespace | Description |
-|  --- | --- |
-|  [Namespace1](./api-documenter-scenarios.namespace1.md) |  |
-
-## Variables
-
-|  Variable | Description |
-|  --- | --- |
-|  [ClassLikeVariable](./api-documenter-scenarios.classlikevariable.md) | Some class-like variable. |
-
-## Type Aliases
-
-|  Type Alias | Description |
-|  --- | --- |
-|  [IInterfaceLikeTypeAlias](./api-documenter-scenarios.iinterfaceliketypealias.md) | Some interface-like type alias. |
-
