--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.namespace1.class3.md
@@ -1,27 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [Namespace1](./api-documenter-scenarios.namespace1.md) &gt; [Class3](./api-documenter-scenarios.namespace1.class3.md)
-
-## Namespace1.Class3 class
-
-
-<b>Signature:</b>
-
-```typescript
-class Class3 
-```
-
-## Properties
-
-|  Property | Modifiers | Type | Description |
-|  --- | --- | --- | --- |
-|  [someProp](./api-documenter-scenarios.namespace1.class3.someprop.md) |  | number | Some prop. |
-
-## Methods
-
-|  Method | Modifiers | Description |
-|  --- | --- | --- |
-|  [someMethod(x)](./api-documenter-scenarios.namespace1.class3.somemethod.md) |  | Some method. |
-|  [someOverload(x)](./api-documenter-scenarios.namespace1.class3.someoverload.md) |  | Some overload. |
-|  [someOverload(x)](./api-documenter-scenarios.namespace1.class3.someoverload_1.md) |  | Some overload. |
-
