--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.namespace1.class3.somemethod.md
@@ -1,24 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [Namespace1](./api-documenter-scenarios.namespace1.md) &gt; [Class3](./api-documenter-scenarios.namespace1.class3.md) &gt; [someMethod](./api-documenter-scenarios.namespace1.class3.somemethod.md)
-
-## Namespace1.Class3.someMethod() method
-
-Some method.
-
-<b>Signature:</b>
-
-```typescript
-someMethod(x: boolean | string): void;
-```
-
-## Parameters
-
-|  Parameter | Type | Description |
-|  --- | --- | --- |
-|  x | boolean \| string |  |
-
-<b>Returns:</b>
-
-void
-
