--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.namespace1.class3.someoverload.md
@@ -1,24 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [Namespace1](./api-documenter-scenarios.namespace1.md) &gt; [Class3](./api-documenter-scenarios.namespace1.class3.md) &gt; [someOverload](./api-documenter-scenarios.namespace1.class3.someoverload.md)
-
-## Namespace1.Class3.someOverload() method
-
-Some overload.
-
-<b>Signature:</b>
-
-```typescript
-someOverload(x: boolean): void;
-```
-
-## Parameters
-
-|  Parameter | Type | Description |
-|  --- | --- | --- |
-|  x | boolean |  |
-
-<b>Returns:</b>
-
-void
-
