--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.namespace1.class3.someprop.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [Namespace1](./api-documenter-scenarios.namespace1.md) &gt; [Class3](./api-documenter-scenarios.namespace1.class3.md) &gt; [someProp](./api-documenter-scenarios.namespace1.class3.someprop.md)
-
-## Namespace1.Class3.someProp property
-
-Some prop.
-
-<b>Signature:</b>
-
-```typescript
-someProp: number;
-```
