--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/api-documenter-scenarios.namespace1.md
@@ -1,19 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [api-documenter-scenarios](./api-documenter-scenarios.md) &gt; [Namespace1](./api-documenter-scenarios.namespace1.md)
-
-## Namespace1 namespace
-
-
-<b>Signature:</b>
-
-```typescript
-export declare namespace Namespace1 
-```
-
-## Classes
-
-|  Class | Description |
-|  --- | --- |
-|  [Class3](./api-documenter-scenarios.namespace1.class3.md) |  |
-
