--- build-tests/api-documenter-scenarios/etc/inheritedMembers/markdown/index.md
@@ -1,12 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md)
-
-## API Reference
-
-## Packages
-
-|  Package | Description |
-|  --- | --- |
-|  [api-documenter-scenarios](./api-documenter-scenarios.md) |  |
-
