+++ common/changes/@microsoft/webpack5-loader-load-themed-styles/issue-3863_2023-01-06-00-24.json
@@ -1,10 +1,10 @@
 {
   "changes": [
     {
-      "packageName": "@microsoft/webpack5-loader-load-themed-styles",
+      "packageName": "@microsoft/webpack5-load-themed-styles-loader",
       "comment": "Introduce webpack5-loader-load-themed-styles",
       "type": "minor"
     }
   ],
-  "packageName": "@microsoft/webpack5-loader-load-themed-styles"
+  "packageName": "@microsoft/webpack5-load-themed-styles-loader"
 }
