+++ common/reviews/api/webpack5-load-themed-styles-loader.api.md
@@ -1,0 +1,19 @@
+## API Report File for "@microsoft/webpack5-load-themed-styles-loader"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+import type { PitchLoaderDefinitionFunction } from 'webpack';
+
+// @public
+export interface ILoadThemedStylesLoaderOptions {
+    async?: boolean;
+    // (undocumented)
+    loadThemedStylesPath?: string;
+}
+
+// @public
+export const pitch: PitchLoaderDefinitionFunction<ILoadThemedStylesLoaderOptions>;
+
+```
