+++ webpack/webpack5-load-themed-styles-loader/LICENSE
@@ -1,4 +1,4 @@
-@microsoft/webpack5-loader-load-themed-styles
+@microsoft/webpack5-load-themed-styles-loader
 
 Copyright (c) Microsoft Corporation. All rights reserved.
