+++ webpack/webpack5-load-themed-styles-loader/src/test/testData/MockStyle1.css
