+++ webpack/webpack5-load-themed-styles-loader/src/test/testData/MockStyle2.ts
