+++ webpack/webpack5-load-themed-styles-loader/tsconfig.json
@@ -2,6 +2,5 @@
   "extends": "./node_modules/@rushstack/heft-node-rig/profiles/default/tsconfig-base.json",
   "compilerOptions": {
     "types": ["heft-jest", "node"]
-  },
-  "include": ["src", "loader-utils.d.ts"]
+  }
 }
