--- webpack/webpack5-loader-load-themed-styles/src/index.ts
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
-// See LICENSE in the project root for license information.
-
-import { LoadThemedStylesLoader } from './LoadThemedStylesLoader';
-
-export default LoadThemedStylesLoader;
